添加api注释 修护swagger不能访问的问题

+++ src/main/java/com/mebay/mapper/UserMapper.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public interface UserMapper {
 
-    User getUserById(@Param("id") int id);
+    User getUserById(@Param("id") Long id);
 
     User getUserByToken(@Param("token") String token);
